@@ -51,12 +51,19 @@ module JdRubyRails
     # This will create an empty whitelist of attributes available for mass-assignment for all models
     # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
     # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = true
+    # config.active_record.whitelist_attributes = true
 
     # Enable the asset pipeline
     config.assets.enabled = true
 
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '1.0'
+
+    # ActiveRecord::Base.establish_connection(
+    #   :adapter => "mysql2",
+    #   :host => "localhost",
+    #   :username => "test",
+    #   :password => "123456",
+    #   :database => "jae-hellofux-development")
   end
 end
